// src/services/openapi/bucket-service/api.ts
export interface BucketCreateRequestDto {
    /**
     * 
     * @type {string}
     * @memberof BucketCreateRequestDto
     */
    module?: string;
    /**
     * 
     * @type {string}
     * @memberof BucketCreateRequestDto
     */
    process?: string;
    /**
     * 
     * @type {string}
     * @memberof BucketCreateRequestDto
     */
    debtorId?: string;
    /**
     * 
     * @type {string}
     * @memberof BucketCreateRequestDto
     */
    bucketProcessId?: string;
    /**
     * 
     * @type {string}
     * @memberof BucketCreateRequestDto
     */
    comment?: string;
    /**
     * 
     * @type {object}
     * @memberof BucketCreateRequestDto
     */
    additionalData?: object;
    /**
     * 
     * @type {boolean}
     * @memberof BucketCreateRequestDto
     */
    syncWithLatestSubmission?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BucketCreateRequestDto
     */
    isGroup?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof BucketCreateRequestDto
     */
    isRelatedToSmi?: boolean;
    /**
     * 
     * @type {string}
     * @memberof BucketCreateRequestDto
     */
    debtorType?: string;
    /**
     * 
     * @type {string}
     * @memberof BucketCreateRequestDto
     */
    debtorRating?: string;
}
/**
 * 
 * @export
 * @interface BucketCreateResponseDto
 */
export interface BucketCreateResponseDto {
    /**
     * 
     * @type {string}
     * @memberof BucketCreateResponseDto
     */
    module?: string;
    /**
     * 
     * @type {string}
     * @memberof BucketCreateResponseDto
     */
    process?: string;
    /**
     * 
     * @type {string}
     * @memberof BucketCreateResponseDto
     */
    debtorId?: string;
    /**
     * 
     * @type {string}
     * @memberof BucketCreateResponseDto
     */
    bucketMasterId?: string;
    /**
     * 
     * @type {string}
     * @memberof BucketCreateResponseDto
     */
    bucketProcessId?: string;
    /**
     * 
     * @type {object}
     * @memberof BucketCreateResponseDto
     */
    additionalData?: object;
}
// ...
export interface GenericSingleDtoBucketCreateResponseDto {
    /**
     * 
     * @type {BucketCreateResponseDto}
     * @memberof GenericSingleDtoBucketCreateResponseDto
     */
    content?: BucketCreateResponseDto;
}