// src/services/openapi/processor-service/api.ts
export interface DynamicStepRequestDto {
    /**
     * 
     * @type {number}
     * @memberof DynamicStepRequestDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof DynamicStepRequestDto
     */
    bucketProcessId?: string;
    /**
     * 
     * @type {string}
     * @memberof DynamicStepRequestDto
     */
    module?: string;
    /**
     * 
     * @type {string}
     * @memberof DynamicStepRequestDto
     */
    process?: string;
    /**
     * 
     * @type {string}
     * @memberof DynamicStepRequestDto
     */
    label?: string;
    /**
     * 
     * @type {string}
     * @memberof DynamicStepRequestDto
     */
    key?: string;
    /**
     * 
     * @type {string}
     * @memberof DynamicStepRequestDto
     */
    urlPath?: string;
    /**
     * 
     * @type {number}
     * @memberof DynamicStepRequestDto
     */
    sequence?: number;
    /**
     * 
     * @type {string}
     * @memberof DynamicStepRequestDto
     */
    roleAccess?: string;
    /**
     * 
     * @type {string}
     * @memberof DynamicStepRequestDto
     */
    statusAccess?: string;
    /**
     * 
     * @type {string}
     * @memberof DynamicStepRequestDto
     */
    statusShow?: string;
    /**
     * 
     * @type {string}
     * @memberof DynamicStepRequestDto
     */
    action?: string;
    /**
     * 
     * @type {string}
     * @memberof DynamicStepRequestDto
     */
    typeStep?: string;
}
// ...
export interface MappingDynamicStepRequestDto {
    /**
     * 
     * @type {string}
     * @memberof MappingDynamicStepRequestDto
     */
    bucketProcessId?: string;
    /**
     * 
     * @type {string}
     * @memberof MappingDynamicStepRequestDto
     */
    module?: string;
    /**
     * 
     * @type {string}
     * @memberof MappingDynamicStepRequestDto
     */
    process?: string;
    /**
     * 
     * @type {string}
     * @memberof MappingDynamicStepRequestDto
     */
    typeStep?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof MappingDynamicStepRequestDto
     */
    keys?: Array<string>;
}
// ...
export class DynamicStepperControllerApi extends BaseAPI {
    /**
     * 
     * @summary API to add stepper
     * @param {DynamicStepRequestDto} dynamicStepRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DynamicStepperControllerApi
     */
    public addStepper(dynamicStepRequestDto: DynamicStepRequestDto, options?: any) {
        return DynamicStepperControllerApiFp(this.configuration).addStepper(dynamicStepRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to create mapping stepper
     * @param {MappingDynamicStepRequestDto} mappingDynamicStepRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DynamicStepperControllerApi
     */
    public createMappingStepper(mappingDynamicStepRequestDto: MappingDynamicStepRequestDto, options?: any) {
        return DynamicStepperControllerApiFp(this.configuration).createMappingStepper(mappingDynamicStepRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to delete mapping stepper
     * @param {DynamicStepRequestDto} dynamicStepRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DynamicStepperControllerApi
     */
    public deleteMappingStepper(dynamicStepRequestDto: DynamicStepRequestDto, options?: any) {
        return DynamicStepperControllerApiFp(this.configuration).deleteMappingStepper(dynamicStepRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to delete stepper
     * @param {DynamicStepRequestDto} dynamicStepRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DynamicStepperControllerApi
     */
    public deleteStepper(dynamicStepRequestDto: DynamicStepRequestDto, options?: any) {
        return DynamicStepperControllerApiFp(this.configuration).deleteStepper(dynamicStepRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to get all dynamic stepper by module and process
     * @param {DynamicStepRequestDto} dynamicStepRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DynamicStepperControllerApi
     */
    public getAllDynamicStepper(dynamicStepRequestDto: DynamicStepRequestDto, options?: any) {
        return DynamicStepperControllerApiFp(this.configuration).getAllDynamicStepper(dynamicStepRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to get stepper by bucket process id
     * @param {DynamicStepRequestDto} dynamicStepRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DynamicStepperControllerApi
     */
    public getStepperBucketProcess(dynamicStepRequestDto: DynamicStepRequestDto, options?: any) {
        return DynamicStepperControllerApiFp(this.configuration).getStepperBucketProcess(dynamicStepRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to get stepper by id
     * @param {DynamicStepRequestDto} dynamicStepRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DynamicStepperControllerApi
     */
    public getStepperById(dynamicStepRequestDto: DynamicStepRequestDto, options?: any) {
        return DynamicStepperControllerApiFp(this.configuration).getStepperById(dynamicStepRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

    /**
     * 
     * @summary API to update stepper
     * @param {DynamicStepRequestDto} dynamicStepRequestDto 
     * @param {*} [options] Override http request option.
     * @throws {RequiredError}
     * @memberof DynamicStepperControllerApi
     */
    public updateStepper(dynamicStepRequestDto: DynamicStepRequestDto, options?: any) {
        return DynamicStepperControllerApiFp(this.configuration).updateStepper(dynamicStepRequestDto, options).then((request) => request(this.axios, this.basePath));
    }

}